Add files via upload
This is the "Master" dataset that includes data on all campaigns on the Karzar website. It contains information on both active and archived campaigns, with their corresponding details such as the campaign URL, start and end dates, campaign title, organizers, province, and category. Please note that this dataset does not include the content or the list of signatories for each campaign.